Broken down in multiple files, added unit tests
