+++ .eslintrc.json
@@ -18,6 +18,7 @@
     "comma-dangle": ["error", "never"],
     "max-len": ["error", { "code": 125, "ignoreUrls": true }],
     "semi": ["error", "always"],
+    "import/prefer-default-export": "off",
     "mocha/no-mocha-arrows": "off",
     "mocha/no-setup-in-describe": "off"
   }
