+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
-  "require": "@babel/register",
+  "require": "@babel/register, @babel/polyfill",
   "extension": ["js"],
   "exit": true,
   "spec": "test/**/*.test.js"
