+++ package.json
@@ -18,6 +18,7 @@
     "http",
     "https",
     "http-transport",
+    "http-transport-emf-stats",
     "rest",
     "aws",
     "emf",
@@ -47,6 +48,7 @@
   "homepage": "https://github.com/bbc/http-transport-emf-stats#readme",
   "devDependencies": {
     "@babel/core": "^7.10.5",
+    "@babel/polyfill": "^7.10.4",
     "@babel/preset-env": "^7.10.4",
     "@babel/register": "^7.10.5",
     "@rollup/plugin-json": "^4.1.0",
@@ -58,7 +60,8 @@
     "mocha": "^8.0.1",
     "nyc": "^15.1.0",
     "rollup": "^2.21.0",
-    "rollup-plugin-terser": "^6.1.0"
+    "rollup-plugin-terser": "^6.1.0",
+    "sinon": "^9.0.2"
   },
   "peerDependencies": {
     "@bbc/http-transport": "3.x"
