+++ src/consts.js
@@ -1,0 +1,11 @@
+export const HTTP_5XX_RESPONSE = 'HTTP_5XX_RESPONSE';
+
+export const actions = [
+  'hit',
+  'miss',
+  'stale',
+  'error',
+  'timeout',
+  'revalidate',
+  'revalidate.error'
+];
